Add code review comment with code suggestion POC
